backend/requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Post-Quantum Cryptography (PQC)
liboqs-python==0.8.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Quantum ML - MUST BE INSTALLED IN THIS ORDER
pennylane==0.33.1
pennylane-qiskit==0.33.1
autoray==0.6.7

# Scientific Computing - MUST use numpy < 1.24 for PennyLane
numpy==1.23.5
scipy==1.11.4
matplotlib==3.8.2

# Computer Vision - OpenCV
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
pillow==10.1.0

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# Deep Learning - PyTorch (compatible with numpy 1.23.5)
torch==2.0.1
torchvision==0.15.2

# Cloud Storage
cloudinary==1.36.0
aiofiles==23.2.1

# HTTP Client
httpx==0.25.1

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1